// Frontend/src/app/models/dashboard.model.ts
export interface ThongKe {
    thangNam: string;
    doanhThu: number;

    tenDSP: string;
    soLuongSanPhamBanRa: number;

    nam: string;
    doanhThuNam: number;
    slSanPhamBanRa: number;

}

export interface ThongKeSoLuong {
    soLuongSP: number;
    soLuongKH: number;
    soLuongDHDaHoanThanh: number;
    soLuongDHChuaXacThuc: number;
    doanhSo: number;
    namHienTai: number;
}